split code gen test files by architecture
