+++ dlk/tests/test_code_generation_base.py
@@ -105,59 +105,6 @@ def get_configurations_by_architecture(test_cases, cpu_name):
     return configurations
 
 
-def get_configurations_x86_64():
-    cpu_name = "x86_64"
-    test_cases = [
-        {'use_avx': True, 'hard_quantize': True, 'threshold_skipping': True},
-        {'use_avx': True, 'hard_quantize': True, 'threshold_skipping': False},
-        {'use_avx': True, 'hard_quantize': False, 'threshold_skipping': False},
-        {'use_avx': False, 'hard_quantize': True, 'threshold_skipping': True},
-        {'use_avx': False, 'hard_quantize': True, 'threshold_skipping': False},
-        {'use_avx': False, 'hard_quantize': False, 'threshold_skipping': False},
-    ]
-    configurations = get_configurations_by_architecture(test_cases, cpu_name)
-
-    additional_test_configuration = updated_dict(dict_codegen_classification(cpu_name),
-                                                 {'use_run_test_script': True})
-    additional_test_cases = [
-        {'use_avx': True, 'input_name': 'raw_image.png', 'test_level': TEST_LEVEL_FUTURE_TARGET},
-        {'use_avx': True, 'input_name': 'preprocessed_image.npy', 'from_npy': True},
-        {'use_avx': False, 'input_name': 'raw_image.png', 'test_level': TEST_LEVEL_FUTURE_TARGET},
-        {'use_avx': False, 'input_name': 'preprocessed_image.npy', 'from_npy': True},
-    ]
-    configurations.extend(get_configurations_by_test_cases(additional_test_cases, additional_test_configuration))
-
-    return [(i, configuration) for i, configuration in enumerate(configurations)]
-
-
-def get_configurations_arm():
-    cpu_name = "arm"
-    test_cases = [
-        {'need_arm_compiler': True, 'hard_quantize': True, 'threshold_skipping': True},
-        {'need_arm_compiler': True, 'hard_quantize': True, 'threshold_skipping': False},
-        # FIXME: Cannot run library with hard_quantize=False for segmentation task.
-        # Issue link: https://github.com/blue-oil/blueoil/issues/512
-        # {'need_arm_compiler': True, 'hard_quantize': False, 'threshold_skipping': True},
-        # {'need_arm_compiler': True, 'hard_quantize': False, 'threshold_skipping': False},
-    ]
-    configurations = get_configurations_by_architecture(test_cases, cpu_name)
-
-    return [(i, configuration) for i, configuration in enumerate(configurations)]
-
-
-def get_configurations_arm_fpga():
-    cpu_name = "arm_fpga"
-    test_cases = [
-        {'need_arm_compiler': True, 'cache_dma': True, 'threshold_skipping': True},
-        {'need_arm_compiler': True, 'cache_dma': True, 'threshold_skipping': False},
-        {'need_arm_compiler': True, 'cache_dma': False, 'threshold_skipping': True},
-        {'need_arm_compiler': True, 'cache_dma': False, 'threshold_skipping': False},
-    ]
-    configurations = get_configurations_by_architecture(test_cases, cpu_name)
-
-    return [(i, configuration) for i, configuration in enumerate(configurations)]
-
-
 class TestCodeGenerationBase(TestCaseDLKBase):
     """Test class for code generation testing."""
 
@@ -379,30 +326,6 @@ class TestCodeGenerationBase(TestCaseDLKBase):
         print(f"Codegen test {prefix}: passed!  {100.0 - percent_failed:.3f}% "
               f"of the output values are correct\n"
               f"[hard quantize == {hard_quantize}, threshold skipping == {threshold_skipping}, cache == {cache_dma}]")
-
-
-class TestCodeGenerationX8664(TestCodeGenerationBase):
-    """Test class for code generation testing."""
-
-    @params(*get_configurations_x86_64())
-    def test_code_generation(self, i, configuration) -> None:
-        self.run_test_all_configuration(i, configuration)
-
-
-class TestCodeGenerationArm(TestCodeGenerationBase):
-    """Test class for code generation testing."""
-
-    @params(*get_configurations_arm())
-    def test_code_generation(self, i, configuration) -> None:
-        self.run_test_all_configuration(i, configuration)
-
-
-class TestCodeGenerationArmFpga(TestCodeGenerationBase):
-    """Test class for code generation testing."""
-
-    @params(*get_configurations_arm_fpga())
-    def test_code_generation(self, i, configuration) -> None:
-        self.run_test_all_configuration(i, configuration)
 
 
 if __name__ == '__main__':
